add dummy KCM for KIC
I'm actually not quite sure what a modern KCM should do vis a vis metadata
so I'm embedding the desktopfile as json and install the desktop file.
seems a bit duplicative though.

also, doesn't do anything yet. smartmonitor needs the notification logic
split into a higher level class so "monitoring" as a visualization
independent use case can exist.
or maybe the kded should expose the data on dbus for consumption
in the KCM. probably more code because that'd be server-client but using
the lib inside the KCM is a waste of computing power and also opens the
door for state inconsistencies between the two (e.g. kded thinks a dev
is kaput but the KCM then reports it working).
'twould also make it easier to write additional UIs (e.g. a plasmoid)

the actual UI plan is to have the notification have a button that pops
up KIC and since we have lots of space there we can give more context on
what's up, what to do, and an array of further apps a user might run
to further inspect or deal with the problem (e.g. run backup app or partman)