refine qml state representation
new model property waiting tracks whether we are currently waiting for
a getmanagedobjects call.

also model this properly on the qml side now using PlaceholderMessages
if the model is invalid (i.e. kded isn't there) display an error.
if the model is valid and not waiting and has no entries display that there
are no smart devices.
message behavior borrowed from discover

there may be use for having a loading message but TBH loading should be
(reasonably) fast so long as kded is already running which it will in
all plasma scenarios