replace open-fstat dance with lstat
since we know the path cannot be relative because of the dir separator
check there is no point in explicitly opening the path when we can
instead pass it to lstat as-is. the path will be absolutely of the form
/dev/foo so at most it could be a symlink that we'll not follow with
lstat.


(cherry picked from commit 00659ebef1f1c741acf989afeb1eac9d4476ef2f)