always try to notify on device appearance
since a hotfix that changed when exactly devices appear the
notifications have been broken as they'd never get emitted when the
failed property was already true when the deviceAdded signal fired (i.e.
pretty much always unless the state changed while the system was
running)

to fix this always try to notify on devices when they appear. if they
aren't failing we aren't raising notifications anyway