report smartctl failure codes to the user
as it turns out there are a number of issues that do not result in an
actually bad SMART status and would not get reported to the user. to
keep the code lean we'll simply look at the exit code bits of smartctl
as they more or less cover all failure scenarios anyway. we'll then
translate those to pretty strings and expose them on the Device objects
for consumption in the qml KCM as "instabilities". they are set visually
apart from actual bad status through different icon and description
because it's hard to say if an instability is in fact indicative of
imminent hard-failure or merely a hiccup (e.g. power loss during a disk
operation).

BUG: 429804
FIXED-in: 5.22