consider devices invalid (ignore them) that have smart disabled
when SMART is disabled smartctl will exit0 but contain no smart_status
which means default values get used and the device is considered
invalid. since we don't know what the status is in this situation we'll
consider the entire data set invalid and thereby hide the device.

with the growing amount of potentially discarded (but present) data
blobs it might soon make sense to think about rendering invalid devices
in the KCM with an unknown state so people at least can find out that
their drives aren't being monitored

BUG: 435699
FIXED-IN: 5.22.0