don't notify on instabilities
this feature was added with the intent to see whether it would raise too
many false positives. unfortunately that didn't turn up during beta
testing but since release many people found this confusing. let's
disable it for 5.22+ again until we can create a better solution.

the instabilities are still shown in the UI. they are not notified on
though

CCBUG: 438539