add magnificent advanced data
to allow users to get more detailed smartctl reports we'll now also
fetch the cli output in addition to the json output. this allows us to
show the verbatim output without much hassle to give knowledge users the
necessary info to actually figure out what exactly is wrong. also this
using a textarea so it may also be easily copied by not advanced users
and pasted on forums or other support channels to ask for help

BUG: 440486